# api/src/doc/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Dancer": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"first_name_en": {
						"type": "string"
					},
					"last_name_en": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image_url": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"first_name_en",
					"last_name_en",
					"created_at",
					"updated_at",
					"description",
					"image_url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Dancer.first_name_en-or-last_name_en-or-description-or-image_url_": {
				"properties": {
					"first_name_en": {
						"type": "string"
					},
					"last_name_en": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image_url": {
						"type": "string"
					}
				},
				"required": [
					"first_name_en",
					"last_name_en",
					"description",
					"image_url"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"DancerCreationParams": {
				"$ref": "#/components/schemas/Pick_Dancer.first_name_en-or-last_name_en-or-description-or-image_url_"
			},
			"Performance": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image_url": {
						"type": "string"
					},
					"start_at": {
						"type": "string",
						"format": "date-time"
					},
					"end_at": {
						"type": "string",
						"format": "date-time"
					},
					"theater": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Theater"
							}
						],
						"nullable": true
					},
					"dancers": {
						"items": {
							"$ref": "#/components/schemas/Dancer"
						},
						"type": "array",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"theater",
					"dancers",
					"created_at",
					"updated_at",
					"description",
					"image_url",
					"start_at",
					"end_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Theater": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"performances": {
						"items": {
							"$ref": "#/components/schemas/Performance"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"location",
					"performances"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Performance.title-or-description-or-image_url-or-start_at-or-end_at_": {
				"properties": {
					"description": {
						"type": "string"
					},
					"image_url": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"start_at": {
						"type": "string",
						"format": "date-time"
					},
					"end_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"description",
					"image_url",
					"title",
					"start_at",
					"end_at"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"PerformanceCreationParams": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_Performance.title-or-description-or-image_url-or-start_at-or-end_at_"
					},
					{
						"properties": {
							"theater_id": {
								"type": "number",
								"format": "double"
							},
							"dancer_ids": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							}
						},
						"type": "object"
					}
				]
			},
			"Pick_Theater.name-or-location_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"location": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"location"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"TheaterCreationParams": {
				"$ref": "#/components/schemas/Pick_Theater.name-or-location_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/dancer": {
			"post": {
				"operationId": "CreateDancer",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Dancer"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DancerCreationParams"
							}
						}
					}
				}
			}
		},
		"/dancers": {
			"get": {
				"operationId": "GetDancers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Dancer"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/dancers/{id}": {
			"get": {
				"operationId": "GetDancer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Dancer"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/performance": {
			"post": {
				"operationId": "CreatePerformance",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Performance"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PerformanceCreationParams"
							}
						}
					}
				}
			}
		},
		"/performances": {
			"get": {
				"operationId": "GetPerformances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Performance"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/performances/{id}": {
			"get": {
				"operationId": "GetPerformance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Performance"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/theater": {
			"post": {
				"operationId": "CreateTheater",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Theater"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TheaterCreationParams"
							}
						}
					}
				}
			}
		},
		"/theaters": {
			"get": {
				"operationId": "GetPerformances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Performance"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/theaters/{id}": {
			"get": {
				"operationId": "GetTheater",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Theater"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}